KamaldocV2/backend/pyproject.toml
[project]
name = "kdoc-backend"
version = "2.0.0"
description = "kdoc V2 — Mobile-first document management with Mistral AI"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "aiosqlite>=0.20.0",
    "asyncpg>=0.30.0",
    "alembic>=1.13.3",
    "httpx>=0.27.2",
    "pyjwt[crypto]>=2.10.0",
    "python-multipart>=0.0.20",
    "boto3>=1.35.0",
    "redis>=5.2.0",
    "arq>=0.26.0",
    "structlog>=24.4.0",
    "stripe>=11.0.0",
    "cryptography>=43.0.0",
    "pymupdf>=1.24.0",
    "pillow>=11.0.0",
    "sentry-sdk[fastapi]>=2.18.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.7.4",
    "mypy>=1.13.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "ASYNC"]
ignore = ["B008"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
